+++ packages/f-content-cards/src/components/ContentCards.vue
@@ -7,7 +7,7 @@
             v-for="(contentCard, cardIndex) in cards">
             <component
                 :is="handleCustomCardType(contentCard.type)"
-                :key="cardIndex"
+                :key="`${cardIndex}_${contentCard.id}`"
                 :card="contentCard"
                 :title="title"
                 :data-test-id="testIdForItemWithIndex(cardIndex)"
@@ -21,9 +21,11 @@
 </template>
 
 <script>
+import { globalisationServices } from '@justeat/f-services';
 import initialiseBraze, { logCardClick, logCardImpressions } from '@justeat/f-metadata';
 import ContentCards from '../services/contentCard.service';
 import cardTemplates from './cardTemplates';
+import tenantConfigs from '../tenants';
 
 /**
  * Generates card-specific analytics data suitable for sending back to GTM via f-trak
@@ -54,47 +56,68 @@ const createBrazeCardEvent = (contentAction, card) => {
 
 export default {
     name: 'ContentCards',
+
     components: {
         ...cardTemplates
     },
+
     props: {
         apiKey: {
             type: String,
             default: ''
         },
-        userId: {
-            type: String,
-            default: ''
+
+        cardLimit: {
+            type: Number,
+            default: -1
         },
-        title: {
-            type: String,
-            default: ''
-        },
+
         enabledCardTypes: {
             type: Array,
             default: () => ([])
         },
+
+        locale: {
+            type: String,
+            default: ''
+        },
+
         pushToDataLayer: {
             type: Function,
             default: () => (() => {})
         },
+
+        showLoadingState: {
+            type: Boolean,
+            default: true
+        },
+
         testId: {
             type: String,
             default: null
         },
-        showLoadingState: {
-            type: Boolean,
-            default: true
+
+        title: {
+            type: String,
+            default: ''
+        },
+
+        userId: {
+            type: String,
+            default: ''
         }
     },
 
     data () {
+        const locale = globalisationServices.getLocale(tenantConfigs, this.locale, this.$i18n);
+        const localeConfig = tenantConfigs[locale];
         const isPostOrderSkeletonCard = this.enabledCardTypes.length && this.enabledCardTypes.every(type => type === 'Post_Order_Card_1');
         const loadingCard = isPostOrderSkeletonCard ? { type: 'postOrder', count: 1 } : { type: 'promo', count: 3 };
 
         return {
             cards: [],
             rawCards: [],
+            copy: { ...localeConfig },
             titleCard: {},
             hasLoaded: false,
             loadingCard
@@ -117,26 +140,39 @@ export default {
         const component = this;
 
         return {
-            emitCardView (card) {
-                component.trackCardVisibility(card);
-            },
             emitCardClick (card) {
                 component.trackCardClick(card);
                 logCardClick(card);
-            }
+            },
+
+            emitCardView (card) {
+                component.trackCardVisibility(card);
+            },
+
+            emitVoucherCodeClick (url) {
+                component.$emit('voucherCodeClick', {
+                    url
+                });
+            },
+
+            copy: component.copy
         };
     },
 
     methods: {
         setupBraze (apiKey, userId, enableLogging = false) {
-            initialiseBraze({
-                apiKey,
-                userId,
-                enableLogging,
-                callbacks: {
-                    handleContentCards: this.contentCards
-                }
-            });
+            try {
+                initialiseBraze({
+                    apiKey,
+                    userId,
+                    enableLogging,
+                    callbacks: {
+                        handleContentCards: this.contentCards
+                    }
+                });
+            } catch (error) {
+                this.$emit('on-error', error);
+            }
         },
 
         contentCards (appboy) {
@@ -148,21 +184,29 @@ export default {
                 .filterCards()
                 .getTitleCard()
                 .arrangeCardsByTitles()
+                .applyCardLimit(this.cardLimit)
                 .output();
 
             this.rawCards = rawCards;
             this.cards = cards;
             this.titleCard = titleCard;
             this.hasLoaded = true;
+
+            this.$emit('on-braze-init', appboy);
+            this.$emit('get-card-count', cards.length);
+            this.$emit('has-loaded', true);
         },
 
         handleCustomCardType (type) {
             switch (type) {
+                case 'Anniversary_Card_1':
+                case 'Voucher_Card_1':
+                    return 'VoucherCard';
+                case 'Post_Order_Card_1':
+                    return 'PostOrderCard';
                 case 'Promotion_Card_1':
                 case 'Promotion_Card_2':
                     return 'PromotionCard';
-                case 'Post_Order_Card_1':
-                    return 'PostOrderCard';
                 default:
                     break;
             }