+++ packages/f-content-cards/src/components/_tests/ContentCards.test.js
@@ -200,7 +200,7 @@ describe('ContentCards', () => {
         const arrange = async () => {
             const PromotionCard = Vue.extend({
                 template: '<div data-promotion-card="true"></div>',
-                inject: ['emitCardClick', 'emitCardView']
+                inject: ['emitCardClick', 'emitCardView', 'emitVoucherCodeClick']
             });
             const cardTypes = ['Promotion_Card_1', 'Promotion_Card_2', 'Post_Order_Card_1'];
             const appboy = createAppboyInstance(cardTypes);
@@ -216,10 +216,16 @@ describe('ContentCards', () => {
             });
             const cardClickHandler = jest.spyOn(instance.vm._provided, 'emitCardClick');
             const cardViewHandler = jest.spyOn(instance.vm._provided, 'emitCardView');
+            const voucherCodeClickHandler = jest.spyOn(instance.vm._provided, 'emitVoucherCodeClick');
             instance.vm.contentCards(appboy);
             await instance.vm.$nextTick();
 
-            return { instance, cardClickHandler, cardViewHandler };
+            return {
+                instance,
+                cardClickHandler,
+                cardViewHandler,
+                voucherCodeClickHandler
+            };
         };
 
         it('should provide a view handler', async () => {
@@ -306,6 +312,32 @@ describe('ContentCards', () => {
                 expect(logCardClick).toHaveBeenCalledWith('foo');
             });
         });
+
+        it('should provide a voucher code copied handler', async () => {
+            const { instance, voucherCodeClickHandler } = await arrange();
+
+            // Act
+            instance.find('[data-promotion-card="true"]').vm.emitVoucherCodeClick(url);
+
+            // Assert
+            expect(voucherCodeClickHandler).toHaveBeenCalledWith(url);
+        });
+
+        describe('the voucher code copied handler', () => {
+            it('should emit a voucherCodeClick event', async () => {
+                const { instance } = await arrange();
+
+                // Act
+                instance.find('[data-promotion-card="true"]').vm.emitVoucherCodeClick(url);
+
+                // Assert
+                expect(instance.emitted().voucherCodeClick).toBeTruthy();
+                expect(instance.emitted().voucherCodeClick.length).toBe(1);
+                expect(instance.emitted().voucherCodeClick[0]).toEqual([{
+                    url
+                }]);
+            });
+        });
     });
 
     describe('when test id prop provided', () => {
@@ -359,5 +391,39 @@ describe('ContentCards', () => {
 
         // Assert
         expect(instance.find('[data-test-id]').exists()).toBeFalsy();
+    });
+
+    describe('emitters', () => {
+        const appboy = createAppboyInstance(['Post_Order_Card_1']);
+
+        const testEmitter = async (emitter, expectedArgs) => {
+            // Arrange & Act
+            const instance = await shallowMount(ContentCards, {
+                propsData: {
+                    apiKey,
+                    userId
+                }
+            });
+            await instance.vm.contentCards(appboy);
+
+            // Assert
+            const [calls] = instance.emitted()[emitter];
+            const [args] = calls;
+
+            expect(calls.length).toBe(1);
+            expect(args).toEqual(expectedArgs);
+        };
+
+        it('should emit an event containing the appboy instance when appboy is initialised', async () => {
+            await testEmitter('on-braze-init', appboy);
+        });
+
+        it('should emit an event containing the content card count when appboy is initialised', async () => {
+            await testEmitter('get-card-count', 1);
+        });
+
+        it('should emit an event containing the loading status when appboy is initialised', async () => {
+            await testEmitter('has-loaded', true);
+        });
     });
 });