+++ packages/f-content-cards/src/components/cardTemplates/index.js
@@ -1,9 +1,11 @@
 import PromotionCard from './PromotionCard.vue';
 import PostOrderCard from './PostOrderCard.vue';
 import SkeletonLoader from './SkeletonLoader.vue';
+import VoucherCard from './VoucherCard.vue';
 
 export default {
     PromotionCard,
     PostOrderCard,
-    SkeletonLoader
+    SkeletonLoader,
+    VoucherCard
 };