+++ packages/f-content-cards/src/tenants/es-ES.js
@@ -1,0 +1,4 @@
+export default {
+    locale: 'es-ES',
+    copyCodeLabel: 'Copia el código'
+};