+++ packages/f-content-cards/stories/ContentCards.stories.js
@@ -2,17 +2,20 @@ import {
     withKnobs,
     optionsKnob as options,
     text,
+    select,
     button
 } from '@storybook/addon-knobs';
+import { action } from '@storybook/addon-actions';
 import { withA11y } from '@storybook/addon-a11y';
 import mock, { proxy } from 'xhr-mock';
 
 import ContentCards from '../src/components/ContentCards.vue';
 import { defaultEnabledCardTypes } from '../src/services/contentCard.service';
 
-import cards from './mockData/cards';
+import cards, { labelledMultiSelectAllowedValues } from './mockData/cards';
 import data from './mockData/data';
 
+const callback = alert;
 /**
  * Resets all locally stored braze data so that the stubbed data is always fresh on page load
  */
@@ -32,22 +35,6 @@ function resetBrazeData () {
         });
 }
 
-const allowedCardTypes = {
-    'Terms and Conditions': 'Terms_And_Conditions_Card',
-    'Terms and Conditions 2': 'Terms_And_Conditions_Card_2',
-    Header: 'Header_Card',
-    Voucher: 'Voucher_Card_1',
-    Recommendation: 'Recommendation_Card_1',
-    'Promotion 1': 'Promotion_Card_1',
-    'Promotion 2': 'Promotion_Card_2',
-    'Home Promotion 1': 'Home_Promotion_Card_1',
-    'Home Promotion 2': 'Home_Promotion_Card_2',
-    'Post Order 1': 'Post_Order_Card_1',
-    'Anniversary 1': 'Anniversary_Card_1',
-    'Restaurant FTC Offer': 'Restaurant_FTC_Offer_Card',
-    'Recommendation 1': 'Recommendation_Card_1'
-};
-
 export default {
     title: 'Components/Organisms',
     decorators: [withKnobs, withA11y]
@@ -63,20 +50,36 @@ export function ContentCardscomponent () {
             apiKey: {
                 default: text('API Key', '00000000-0000-0000-0000-000000000000')
             },
+
             userId: {
                 default: text('User ID', 'test-user-id')
             },
+
             title: {
                 default: text('Title', 'Promotional Offers')
             },
+
+            locale: {
+                default: select('Locale', ['da-DK', 'en-GB', 'en-AU'], 'en-GB')
+            },
+
             enabledCardTypes: {
-                default: options('Enabled Card Types', allowedCardTypes, defaultEnabledCardTypes, {
+                default: options('Enabled Card Types', labelledMultiSelectAllowedValues, defaultEnabledCardTypes, {
                     display: 'multi-select'
                 })
             },
+
             refreshDisplayedCards: {
                 default: button('Refresh Card Types Displayed', () => { window.appboy.requestContentCardsRefresh(); })
             }
+        },
+
+        methods: {
+            onBrazeInit: action('on-braze-init'),
+            getCardCount: action('get-card-count'),
+            getTitleCard: action('get-title-card'),
+            hasLoaded: action('has-loaded'),
+            onError: action('on-error')
         },
 
         /**
@@ -98,9 +101,14 @@ export function ContentCardscomponent () {
         },
 
         template: `<content-cards
+            @on-braze-init="onBrazeInit"
+            @get-card-count="getCardCount"
+            @has-loaded="hasLoaded"
+            @on-error="onError"
             :userId="userId"
             :apiKey="apiKey"
             :title="title"
+            :locale="locale"
             :enabledCardTypes="enabledCardTypes" />`
     };
 }