+++ packages/f-content-cards/stories/mockData/cards.js
@@ -99,7 +99,7 @@ const fieldTypeToFaker = {
  * @type {any}
  */
 export const labelledMultiSelectAllowedValues = Object.fromEntries(Object.entries(cardTypes)
-    .map(([key, { label }]) => [key, label]));
+    .map(([key, { label }]) => [label, key]));
 
 function timeRange10HoursAroundNow () {
     const now = Math.floor(Date.now() / 1000);
@@ -156,7 +156,9 @@ function seededRandomCardOfType (type) {
         e,
         tp: 'short_news',
         ar: 1,
-        i: null,
+        i: (type === 'Anniversary_Card_1'
+            ? `https://picsum.photos/seed/${type}_i/109/96`
+            : `https://picsum.photos/seed/${type}_i/384/216?blur=3`),
         u: null,
         uw: null,
         tt,